requirements.txt
Pillow>=10.0.0
numpy<2,>=1.24.0
easyocr>=1.7.0
torch>=2.0.0
torchvision>=0.15.0
